Improve lexer to also support directives
